+++ src/primaite/simulator/system/services/dns/dns_server.py
@@ -26,6 +26,7 @@ class DNSServer(Service):
         # TCP for now
         kwargs["protocol"] = IPProtocol.TCP
         super().__init__(**kwargs)
+        self.start()
 
     def describe_state(self) -> Dict:
         """