+++ tests/e2e_integration_tests/test_uc2_data_manipulation_scenario.py
@@ -1,7 +1,7 @@
 from primaite.simulator.network.hardware.nodes.computer import Computer
 from primaite.simulator.network.hardware.nodes.server import Server
 from primaite.simulator.system.applications.database_client import DatabaseClient
-from primaite.simulator.system.services.database_service import DatabaseService
+from primaite.simulator.system.services.database.database_service import DatabaseService
 from primaite.simulator.system.services.red_services.data_manipulation_bot import DataManipulationBot